Add code to plot Column Space of A
In case of A being a rectangular matrix of size m*n
A solution to Ax = b, exist for those b which lie in the column space
of A.

The code which I have added in this commit, plots this column space
for m=3 and n=2. This is a subspace in R^3. The example I have taken
from gilbert strang's book turns out to have only 2 pivots (You can try
verifying using rref(A)). Thus We get the column space to be a plane.

Note that [0;0;0], and both columns of A lie in the
subspace/column space. Next I hope to be able to plot a b which doesnt
lie on this plane and show that its projection onto this plane is the
best **approximate** solution to the Ax=b problem.

Signed-off-by: Aditya A Prasad <adityaprasadiskool@gmail.com>